backend/requirements.txt
asgiref==3.6.0
Django==3.2
django-filter==2.4.0
django-colorfield==0.8.0
django-extensions==3.2.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
flake8==5.0.4
gunicorn==20.1.0
isort==5.10.1
django-colorfield==0.8.0
psycopg2-binary==2.9.4
Pillow==9.4.0
PyJWT==2.6.0
pytest-django==3.8.0
python-dotenv==0.20.0
pytz==2020.1
sorl-thumbnail==12.9.0
sqlparse==0.3.1
requests==2.28.2
















